+++ src/edu/jhu/thrax/hadoop/datatypes/FeatureValue.java
@@ -1,6 +1,6 @@
 package edu.jhu.thrax.hadoop.datatypes;
 
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.GenericWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
@@ -9,7 +9,7 @@ import org.apache.hadoop.io.Writable;
 public class FeatureValue extends GenericWritable {
 
   @SuppressWarnings("rawtypes")
-  private static Class[] TYPES = {DoubleWritable.class, IntWritable.class, Text.class,
+  private static Class[] TYPES = {FloatWritable.class, IntWritable.class, Text.class,
       Annotation.class, AlignmentWritable.class};
 
   FeatureValue() {}
