+++ src/edu/jhu/thrax/hadoop/features/annotation/LogCountFeature.java
@@ -2,7 +2,7 @@ package edu.jhu.thrax.hadoop.features.annotation;
 
 import java.util.Set;
 
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.mapreduce.Reducer.Context;
@@ -15,7 +15,7 @@ import edu.jhu.thrax.hadoop.jobs.ThraxJob;
 public class LogCountFeature implements AnnotationFeature {
 
   private static final Text LABEL = new Text("LogCount");
-  private static final DoubleWritable ZERO = new DoubleWritable(0);
+  private static final FloatWritable ZERO = new FloatWritable(0);
 
   public Text getName() {
     return LABEL;
@@ -31,7 +31,7 @@ public class LogCountFeature implements AnnotationFeature {
 
   @Override
   public Writable score(RuleWritable r, Annotation annotation) {
-    return new DoubleWritable(Math.log(annotation.count()));
+    return new FloatWritable((float) Math.log(annotation.count()));
   }
 
   @Override
