+++ src/edu/jhu/thrax/hadoop/features/mapred/SourcePhraseGivenTargetFeature.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import java.util.Arrays;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
@@ -72,7 +72,7 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
 
   private static class Reduce extends Reducer<RuleWritable, IntWritable, RuleWritable, FeaturePair> {
     private int marginal;
-    private DoubleWritable prob;
+    private FloatWritable prob;
     private static final Text NAME = new Text("p(f|e)");
 
     protected void setup(Context context) throws IOException, InterruptedException {
@@ -93,7 +93,7 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
         int count = 0;
         for (IntWritable x : values)
           count += x.get();
-        prob = new DoubleWritable(-Math.log(count / (double) marginal));  
+        prob = new FloatWritable((float) -Math.log(count / (float) marginal));  
         return;
       }
       context.write(key, new FeaturePair(NAME, prob));
@@ -135,7 +135,7 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
   }
 
 
-  private static final DoubleWritable ZERO = new DoubleWritable(0.0);
+  private static final FloatWritable ZERO = new FloatWritable(0.0f);
 
   public void unaryGlueRuleScore(Text nt, java.util.Map<Text, Writable> map) {
     map.put(Reduce.NAME, ZERO);
